rename pkg to trace
