+++ internal/trace/migrate.go
@@ -1,4 +1,4 @@
-package indexer
+package trace
 
 import "database/sql"
