+++ internal/trace/migrate_test.go
@@ -1,4 +1,4 @@
-package indexer
+package trace
 
 import (
 	"testing"
