+++ internal/trace/sql.go
@@ -1,4 +1,4 @@
-package indexer
+package trace
 
 import (
 	_ "modernc.org/sqlite"
