+++ internal/trace/test_case.go
@@ -1,4 +1,4 @@
-package indexer
+package trace
 
 import (
 	"context"
